+++ packages/shared-rtc-bench/tests/topology-delivery/rtc-topology-delivery-log-performance-harness.test.ts
@@ -4,6 +4,7 @@ import {
   RTC_TOPOLOGY_DELIVERY_LOG_BENCHMARK_POLICY,
   summarizeRtcTopologyDeliveryLatencies,
 } from '../../topology-delivery/delivery-log-bench.ts';
+import { registerRtcTopologyDeliveryBenchmarkStreams } from '../../topology-delivery/run-rtc-topology-delivery-log-workloads.ts';
 
 describe('RTC topology delivery-log performance harness', () => {
   it('keeps the one-stream and three-stream workloads directly comparable', () => {
@@ -26,5 +27,19 @@ describe('RTC topology delivery-log performance harness', () => {
     expect(() => summarizeRtcTopologyDeliveryLatencies([])).toThrow(
       'RTC topology delivery benchmark latency samples are required',
     );
+  });
+
+  it('cleans only benchmark-owned streams when registration fails partway', async () => {
+    const cleaned: string[][] = [];
+    await expect(
+      registerRtcTopologyDeliveryBenchmarkStreams({
+        streams: ['owned-1', 'collision', 'never-attempted'],
+        register: async (streamId) => (streamId === 'collision' ? 'collision' : 'registered'),
+        cleanup: async (streams) => {
+          cleaned.push([...streams]);
+        },
+      }),
+    ).rejects.toThrow('stream collision: collision');
+    expect(cleaned).toEqual([['owned-1']]);
   });
 });
