+++ packages/shared-rtc-bench/workloads/topology/rtc-topology-inactive-churn-bench.ts
@@ -1,5 +1,7 @@
 import type { GroupSnapshot } from '@shared/api/group-types.ts';
-import { RallarRtcTopologyService } from '@shared-server/rallar-system/services/rallar-rtc-topology-service.ts';
+// prettier-ignore
+import { RallarRtcTopologyService } from
+  '@shared-server/rallar-system/services/rallar-rtc-topology-service.ts';
 // prettier-ignore
 import { createDeterministicRtcTopologyGroupSnapshot } from
   './create-deterministic-rtc-topology-group-snapshot.ts';
@@ -15,6 +17,9 @@ import {
   parseRtcBaselineOneTokenOptions,
 } from '../../baseline/command/rtc-baseline-cli-options.ts';
 import { validateRtcBaselineId } from '../../baseline/contracts/rtc-baseline-validation.ts';
+// prettier-ignore
+import { runRtcBaselineAcceptedWorkerSamples } from
+  '../../baseline/acceptance/rtc-baseline-failure-accounting.ts';
 
 export type RtcTopologyInactiveChurnMode = 'retain' | 'cleanup';
 interface RtcTopologyInactiveChurnInput {
@@ -107,26 +112,17 @@ export async function runRtcTopologyInactiveChurnAcceptedSamples(input: {
   readonly worker: RtcTopologyInactiveChurnAcceptedArguments;
   readonly run: () => Promise<RtcTopologyInactiveChurnResult> | RtcTopologyInactiveChurnResult;
 }): Promise<RtcBaselineSampleDto[]> {
-  const samples: RtcBaselineSampleDto[] = [];
-  let failureId: string | undefined;
-  for (let index = 0; index < input.worker.sampleIds.length; index += 1) {
-    const identity = createIdentity(input.worker, index);
-    if (failureId !== undefined) {
-      samples.push(
-        createSample(identity, null, [
-          rtcBaselineIssue('$.rawEvidence', 'causal-not-run', failureId),
-        ]),
-      );
-      continue;
-    }
-    const result = await input.run();
-    const issues = validateResult(input.worker.input, result);
-    if (issues.length > 0) {
-      failureId = identity.sampleId;
-    }
-    samples.push(createSample(identity, result, issues));
-  }
-  return samples;
+  return runRtcBaselineAcceptedWorkerSamples({
+    worker: {
+      ...input.worker,
+      workloadId: 'RTC-B03',
+      caseId: 'topology-inactive-churn',
+      inputKey: `mode-${input.worker.input.mode}`,
+    },
+    run: input.run,
+    validate: (result) => validateResult(input.worker.input, result),
+    createSample: ({ identity, result, issues }) => createSample(identity, result, issues),
+  });
 }
 
 function parseDiagnosticArguments(options: Readonly<Record<string, string>>) {
@@ -234,21 +230,6 @@ function createExpectedSampleIds(
     { length: 3 },
     (_value, index) => `${prefix}-${String(index + 1).padStart(3, '0')}`,
   );
-}
-
-function createIdentity(
-  worker: RtcTopologyInactiveChurnAcceptedArguments,
-  index: number,
-): RtcBaselineSampleIdentityDto {
-  return {
-    sampleId: worker.sampleIds[index],
-    workloadId: 'RTC-B03',
-    caseId: 'topology-inactive-churn',
-    inputKey: `mode-${worker.input.mode}`,
-    intendedPhase: worker.intendedPhase,
-    outerOrdinal: worker.outerOrdinal,
-    innerOrdinal: index + 1,
-  };
 }
 
 function createSample(
