+++ packages/shared-rtc-bench/workloads/topology/rtc-topology-tree-no-rtt-bench.ts
@@ -1,4 +1,6 @@
-import { RallarRtcTopologyService } from '@shared-server/rallar-system/services/rallar-rtc-topology-service.ts';
+// prettier-ignore
+import { RallarRtcTopologyService } from
+  '@shared-server/rallar-system/services/rallar-rtc-topology-service.ts';
 // prettier-ignore
 import { createDeterministicRtcTopologyGroupSnapshot } from
   './create-deterministic-rtc-topology-group-snapshot.ts';
@@ -14,6 +16,9 @@ import {
   parseRtcBaselineOneTokenOptions,
 } from '../../baseline/command/rtc-baseline-cli-options.ts';
 import { validateRtcBaselineId } from '../../baseline/contracts/rtc-baseline-validation.ts';
+// prettier-ignore
+import { runRtcBaselineAcceptedWorkerSamples } from
+  '../../baseline/acceptance/rtc-baseline-failure-accounting.ts';
 
 interface RtcTopologyTreeInput {
   readonly sessions: number;
@@ -75,26 +80,17 @@ export async function runRtcTopologyTreeAcceptedSamples(input: {
   readonly worker: RtcTopologyTreeAcceptedArguments;
   readonly run: () => Promise<RtcTopologyTreeResult> | RtcTopologyTreeResult;
 }): Promise<RtcBaselineSampleDto[]> {
-  const samples: RtcBaselineSampleDto[] = [];
-  let failureId: string | undefined;
-  for (let index = 0; index < input.worker.sampleIds.length; index += 1) {
-    const identity = createIdentity(input.worker, index);
-    if (failureId !== undefined) {
-      samples.push(
-        createSample(identity, null, [
-          rtcBaselineIssue('$.rawEvidence', 'causal-not-run', failureId),
-        ]),
-      );
-      continue;
-    }
-    const result = await input.run();
-    const issues = validateResult(input.worker.input, result);
-    if (issues.length > 0) {
-      failureId = identity.sampleId;
-    }
-    samples.push(createSample(identity, result, issues));
-  }
-  return samples;
+  return runRtcBaselineAcceptedWorkerSamples({
+    worker: {
+      ...input.worker,
+      workloadId: 'RTC-B03',
+      caseId: 'topology-tree',
+      inputKey: `sessions-${input.worker.input.sessions}`,
+    },
+    run: input.run,
+    validate: (result) => validateResult(input.worker.input, result),
+    createSample: ({ identity, result, issues }) => createSample(identity, result, issues),
+  });
 }
 
 function parseDiagnosticArguments(options: Readonly<Record<string, string>>) {
@@ -207,21 +203,6 @@ function createExpectedSampleIds(
     { length: 5 },
     (_value, index) => `${prefix}-${String(index + 1).padStart(3, '0')}`,
   );
-}
-
-function createIdentity(
-  worker: RtcTopologyTreeAcceptedArguments,
-  index: number,
-): RtcBaselineSampleIdentityDto {
-  return {
-    sampleId: worker.sampleIds[index],
-    workloadId: 'RTC-B03',
-    caseId: 'topology-tree',
-    inputKey: `sessions-${worker.input.sessions}`,
-    intendedPhase: worker.intendedPhase,
-    outerOrdinal: worker.outerOrdinal,
-    innerOrdinal: index + 1,
-  };
 }
 
 function createSample(
